+++ R/html_dependency.R
@@ -352,10 +352,19 @@ copyDependencyToDir <- function(dependency, outputDir, mustWork = TRUE) {
   # relatively safe to be removed recursively
   if (dir_exists(target_dir)) unlink(target_dir, recursive = TRUE)
   dir.create(target_dir)
+  dependency$src$file <- normalizePath(target_dir, "/", TRUE)
 
   files <- if (dependency$all_files) list.files(dir) else {
     unlist(dependency[c('script', 'stylesheet', 'attachment')])
   }
+
+  if (length(files) == 0) {
+    # This dependency doesn't include any files
+    # no need to copy and we can clean up the target directory
+    unlink(target_dir, recursive = TRUE)
+    return(dependency)
+  }
+
   srcfiles <- file.path(dir, files)
   if (any(!file.exists(srcfiles))) {
     stop(
@@ -376,8 +385,6 @@ copyDependencyToDir <- function(dependency, outputDir, mustWork = TRUE) {
       dir.create(to)
     file.copy(from, to, overwrite = TRUE, recursive = isdir, copy.mode = FALSE)
   }, srcfiles, destfiles, isdir)
-
-  dependency$src$file <- normalizePath(target_dir, "/", TRUE)
 
   dependency
 }