+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Closed #290: htmltools previously did not specify which version of fastmap to use, and would fail to install with an old version of fastmap. (#291)
 
+* `copyDependencyToDir()` no longer creates empty directories for dependencies that do not have any files. (@gadenbuie, #276)
+
 # htmltools 0.5.2
 
 ## Breaking Changes